월별 통계 총금액 타입 수정 - 20210329

+++ src/main/java/com/rest/api/entity/fastcafe_stat/StatVanPayMonthly.java
@@ -25,6 +25,6 @@ public class StatVanPayMonthly {
     private int branchMachineId;
     private String baseYear;
     private String baseMonth;
-    private int payMoney;
+    private double payMoney;
     private int payCnt;
 }
